requirements.txt
torch
torchvision
matplotlib
numpy
Pillow>=10.2.0